Unit Test for EmployeeRepo expanded with TestInitialize and TestCleanup
